+++ private_values/ini.py
@@ -17,10 +17,8 @@ class PrivateIni(PrivateBase):
             ini.read_string(self.filepath.read_text())
         except Exception as exx:
             msg = f"[CRITICAL] incorrect format file!\n{exx!r}"
-            if self.RAISE:
-                raise Exx_PvNotAccepted(msg)
-            else:
-                return {}
+            print(msg)
+            raise exx
 
         if not self.SECTION or self.SECTION == "DEFAULT" or ini.has_section(section=self.SECTION):
             result = dict(ini[self.SECTION or "DEFAULT"])
@@ -28,10 +26,8 @@ class PrivateIni(PrivateBase):
         else:
             msg = f"[CRITICAL] NO [{self.SECTION=} in {self.filepath=}]\n"
             msg += self.filepath.read_text()
-            if self.RAISE:
-                raise Exx_PvNotAccepted(msg)
-            else:
-                return {}
+            print(msg)
+
 
 # =====================================================================================================================
 class PrivateAuthIni(PrivateAuth, PrivateIni):