+++ private_values/json.py
@@ -16,10 +16,7 @@ class PrivateJson(PrivateBase):
         except Exception as exx:
             msg = f"[CRITICAL] incorrect format file!\n{exx!r}"
             print(msg)
-            if self.RAISE:
-                raise Exx_PvNotAccepted(msg)
-            else:
-                return {}
+            raise exx
 
         if self.SECTION:
             json_data = json_data.get(self.SECTION)
@@ -29,10 +26,7 @@ class PrivateJson(PrivateBase):
         else:
             msg = f"[CRITICAL] NO [{self.SECTION=} in {self.filepath=}]\n"
             msg += self.filepath.read_text()
-            if self.RAISE:
-                raise Exx_PvNotAccepted(msg)
-            else:
-                return {}
+            print(msg)
 
 
 # =====================================================================================================================